// types/dashboard.ts
export interface Overview {
    totalMembers: number;
    newMembersThisMonth: number;
    totalCounselors: number;
    newCounselorsThisMonth: number;
    totalBookings: number;
    bookingsThisMonth: number;
    bookingRevenue: number;
    bookingRevenueThisMonth: number;
    totalCoursesPurchased: number;
    coursesPurchasedThisMonth: number;
    courseRevenue: number;
    courseRevenueThisMonth: number;
    totalMemberships: number;
    membershipsThisMonth: number;
    membershipRevenue: number;
    membershipRevenueThisMonth: number;
}